Restructured packages based on io, nutch and spark
